requirements.txt
lpips==0.1.4
numpy==2.1.2
PyYAML==6.0.2
torch==2.5.0
torchattacks==3.5.1
torchvision==0.20.0
